refactor 해더 쿠키 리팩토링
